Implented a simple version system for the pickled data. This means old flax.flx file will (this time) cause an error, but thereafter a warning about incompatible versions will be emmitted and data reset to he defaults.
Moved the log configuration further down the startup code so that the
root logger is configured before we issue any warnings.

Changes for "similar document search" as per
http://code.google.com/p/flaxcode/issues/detail?id=21

Added a link for similar searches, not quite done yet as we need to
get back a string from the query (i.e. implement search.unparse()
properly)