http://code.google.com/p/flaxcode/issues/detail?id=11
Made a start on mapping for paths, the mapping values are not yet used
and we need to inlude some javascript to push extra input boxes into
the page.


FIXES: http://code.google.com/p/flaxcode/issues/detail?id=21

Mappings are now used to generate the urls for files, and title are
used if available, to the work for issue 21 is done. (Probably some
cosmetic work on the "search for similar" link is desirable.)

http://code.google.com/p/flaxcode/issues/detail?id=11

Added code to generate urls from doc ids based on the mappings. Also
fixed a problem with adding lots of blank mapping enties on
round-tripping.

This "works" now, but still could do with some javascript to add extra
path/mapping input boxes without round-tripping.