http://code.google.com/p/flaxcode/issues/detail?id=7
Made infrastructure for discovering status of indexer and used the
results in the collection list page. I'm not exactly sure what
statistics are useful or interesting, but at least now it should be
straigtforward to add things to the dictionaries used for this
information.

We use a search connection for collections in a few places so I've
cached one in the DocCollection object and (ensured that we don't try
to pickle it).