Refactored the way that indexing is controlled. All the logic about what gets indexed and when is in the main process. The indexing process just responds to requests to index things.
The shared state for reporting on the indexing progress and for
prematurely stopping the indexer is all done via shared memory now.

Cached information about the last indexing of document collections in
the collection object itself, so that something useful can be shown
after a restart. Updated the UI to display such information.

Fixed a problem with log configuration propogation to the indexing
process.