http://code.google.com/p/flaxcode/issues/detail?id=14
Set a collection as due for indexing after it's been update.

This change doesn't stop a currently indexing collection as suggested
in comment 3. I'm not sure what's needed to acheive this atm. There's
no queue of actions for each collection, so we'd need to hold it until
it stops, then unhold it. But there are either synchronisation or
blocking issues.