FIXES: http://code.google.com/p/flaxcode/issues/detail?id=14
Added a stop_indexing method to IndexServer and used to stop indexing
after a collection is updated. Added a lock to ensure that changes to
.stop_sv and and currently_indexing happen togther.

In indexer.do_indexing added a flush for the indexer connection in the
main loop. This was because a) Ram usage can get large, and b)
holding/stopping indexing can seem very unresponsive if there is still
a lot of work outstanding when the stop flag is set.