Formed an appropriate query from the advanced search form inputs.
Refactored passing of arguments to search rendering, since advanced is
now implicit in the template.

Added extra data to the key to the template cache to account for the
fact that different urls can be generated from the same file name, so
we need to differentiate the cached templates on this basis.