FIXES: http://code.google.com/p/flaxcode/issues/detail?id=43
Switched to using SyncManager rather than LocalManager to deal with
issue 43. 

I think the route cause of the problem is probably a bug in the
processing LocalManager code (on Windows at least). The cleanup code
for the localmanger tries to delete a temp file whilst another process
still has a handle on it. If you terminate subprocesses manually then
the problem goes away, but the whole point of having daemonic
processes is that the infrastructure is supposed to tidy them up for
you.