http://code.google.com/p/flaxcode/issues/detail?id=109
Added some logging (not that "DEBUG" level for this filter will now
give all the text it's yielding")

Continued through the "unspecified" com errors, since this seems to
work. I've added a check that we don't try and repeat the same chunk
in case this exception is raised in other situations where the same
chunk might be repeated once this exception is raised.